requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
python-telegram-bot>=20.0
pyyaml>=6.0
python-dotenv>=1.0.0
schedule>=1.2.0
fake-useragent>=1.4.0
googlesearch-python>=1.2.0
